% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POD_TPI_df.R
\name{POD_TPI_df}
\alias{POD_TPI_df}
\title{My Exported Function}
\usage{
POD_TPI_df(
  df,
  W,
  MaxDose,
  p_T,
  epsilon = c(0.05, 0.05),
  niter = 1000,
  type_dose_decision = "mTPI",
  type_p_prior = "uniform",
  type_t_model = "pwuniform",
  type_suspension = 2,
  type_safety = 2,
  q1 = 1,
  q2 = 0.15
)
}
\arguments{
\item{x}{A parameter description.}
}
\value{
The result of the function.
}
\description{
This function does something useful.
}
